refactor: conformance adapter and benchmark model
